Refactor: FrontmatterKeyInfo を schema_meta.py に分離 (#178)
stats.py が「時系列集計 (Reindex/VaultStats)」と「静的スキーマ記述
(FrontmatterKeyInfo)」の両性質を抱えていた ambiguity を解消。

- src/vault_search/schema_meta.py: 新規。FrontmatterKeyInfo を移動
- src/vault_search/stats.py: ReindexStats / VaultStats のみ残し、
  docstring を narrower scope に書き換え
- src/vault_search/indexer.py / mcp_contract.py: import 経路を更新
- tests: assertion message から旧モジュール名を削除

後続 PR 1b #38 で追加する ToolOverview / recommended_flow 等の schema メタ
型は schema_meta.py 側に集約する方針。
